+++ template/UtilisateurDetail.php
@@ -18,7 +18,8 @@
  * @var array $tokens
  * @var array $tree
  * @var bool $mfa_enabled
- * @var bool $mfa_admin_disable
+ * @var bool $mfa_admin_reset
+ * @var bool $mfa_obligatory
  */
 
 use Pastell\Utilities\Certificate;
@@ -80,18 +81,35 @@ use Pastell\Utilities\Certificate;
             <td>
                 <?php echo $mfa_enabled ? 'Activée' : 'Désactivée' ?>
                 <?php if ($mfa_enabled && (int)$id_u === (int)$id_current_u) : ?>
-                    <a href='Mfa/authRequired?action=desactivation' class='btn btn-danger btn-sm'>
-                        <i class='fas fa-shield-alt'></i>&nbsp;Désactiver
-                    </a>
-                <?php elseif ($mfa_admin_disable && $mfa_enabled) : ?>
-                    <form action='Utilisateur/disableMfa' method='post' class='d-inline'>
-                        <?php $this->displayCSRFInput() ?>
-                        <input type='hidden' name='id_u' value='<?php echo $id_u ?>'/>
-                        <button type='submit' class='btn btn-danger btn-sm'
-                                onclick="return confirm('Désactiver la double authentification de cet utilisateur ?')">
+                    <?php if (!empty($mfa_obligatory)) : ?>
+                        <a href='Mfa/authRequired?action=reinitialisation' class='btn btn-warning btn-sm'>
+                            <i class='fas fa-rotate'></i>&nbsp;Réinitialiser
+                        </a>
+                    <?php else : ?>
+                        <a href='Mfa/authRequired?action=desactivation' class='btn btn-danger btn-sm'>
                             <i class='fas fa-shield-alt'></i>&nbsp;Désactiver
-                        </button>
-                    </form>
+                        </a>
+                    <?php endif; ?>
+                <?php elseif ($mfa_admin_reset && $mfa_enabled) : ?>
+                    <?php if (!empty($mfa_obligatory)) : ?>
+                        <form action='Utilisateur/resetMfa' method='post' class='d-inline'>
+                            <?php $this->displayCSRFInput() ?>
+                            <input type='hidden' name='id_u' value='<?php echo $id_u ?>'/>
+                            <button type='submit' class='btn btn-warning btn-sm'
+                                    onclick="return confirm('Réinitialiser la double authentification de cet utilisateur ?')">
+                                <i class='fas fa-rotate'></i>&nbsp;Réinitialiser
+                            </button>
+                        </form>
+                    <?php else : ?>
+                        <form action='Utilisateur/disableMfa' method='post' class='d-inline'>
+                            <?php $this->displayCSRFInput() ?>
+                            <input type='hidden' name='id_u' value='<?php echo $id_u ?>'/>
+                            <button type='submit' class='btn btn-danger btn-sm'
+                                    onclick="return confirm('Désactiver la double authentification de cet utilisateur ?')">
+                                <i class='fas fa-shield-alt'></i>&nbsp;Désactiver
+                            </button>
+                        </form>
+                    <?php endif; ?>
                 <?php endif; ?>
             </td>
         </tr>